moved around all n_grams things to its own ngrams.py file in order to make easier and more logical to read the code base, will have to add changes in ngrams.py to allow for the reading and writing of the ngrams globals instead of in globals.py. only ngrams alters and acesses these variables so it's more logical to include them there
